Initial code commit
This deserves a little more fanfare
Simple bot to connect to the twitch irc servers and listen to a channel.
It does some very simple analysis of messages via the top 10000 english
words file gives them a pass/fail.

TODO: Add in the code to call out to the twitch apis anc get list of
emotes as well as the better twitch tv emotes